+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Unreleased
-...
+- **feat**: implement `Display` for `Location`
 
 ## 0.1.1
 *[start of log]*